storage: move snapshot_* methods from controller to collections
Work towards implementing https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/20240117_decoupled_storage_controller.md
